// mods/appmanager/src/service/protos/appmanager_pb.d.ts
    export type AsObject = {
        pageSize: number,
        pageToken: string,
        view: common_pb.View,
    }
// ...
export class App extends jspb.Message { 
    getRef(): string;
    setRef(value: string): App;
    getName(): string;
    setName(value: string): App;
    getDescription(): string;
    setDescription(value: string): App;
    getCreateTime(): string;
    setCreateTime(value: string): App;
    getUpdateTime(): string;
    setUpdateTime(value: string): App;
    getStatus(): App.Status;
    setStatus(value: App.Status): App;
    getAccessKeyId(): string;
    setAccessKeyId(value: string): App;

    getLabelsMap(): jspb.Map<string, string>;
    clearLabelsMap(): void;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): App.AsObject;
    static toObject(includeInstance: boolean, msg: App): App.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: App, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): App;
    static deserializeBinaryFromReader(message: App, reader: jspb.BinaryReader): App;
}

export namespace App {
    export type AsObject = {
        ref: string,
        name: string,
        description: string,
        createTime: string,
        updateTime: string,
        status: App.Status,
        accessKeyId: string,

        labelsMap: Array<[string, string]>,
    }

    export enum Status {
    UNKNOWN = 0,
    CREATING = 1,
    RUNNING = 2,
    STOPPED = 3,
    REMOVED = 4,
    }

}